Correct Output for input_file2.asc
I ran the program using input_file2.asc and managed to confirm that it produces the correct output to register and memory values. That correct output is here in this file I'm committing. I didn't have to make any changes to the program to get it to produce the correct output for this second test case :)